# src/routes/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Consultas de DNI , RUC, TIPO CAMBIO, COMISION PRIMA AFP - **Perú**",
        "version": "1.0.0",
        "title": "Consulta Ruc, Dni, Tipo Cambio, Comisiones Prima SBS - Perú",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Alex Pariona",
            "email": "lx7pary@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "basePath": "/",
...
    "schemes": [
        "https",
        "http"
    ],
    "paths": {
...
        "/public/api/tc-sunat/actual": {
            "get": {
                "tags": [
                    "SUNAT"
                ],
                "summary": "TIPO CAMBIO",
                "description": "Retorna tipo de cambio del dia",
                "operationId": "getTipoCambio",
                "produces": [
                    "application/json"
                ],
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Operacion exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/definitions/TipoCambioResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tasa Cambio invalido o no encontrado"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "comision_auth": {
            "type": "oauth2",
            "authorizationUrl": "https://consultape.swagger.io/oauth/dialog",
            "flow": "implicit",
            "scopes": {
                "write:pets": "modify searchpe in your account",
                "read:pets": "read your searchpe"
            }
        },
        "api_key": {
            "type": "apiKey",
            "name": "api_key",
            "in": "header"
        }
    },
    "definitions": {
...
        "TipoCambioResponse": {
            "type": "object",
            "properties": {
                "dia": {
                    "type": "integer"
                },
                "compra": {
                    "type": "integer",
                    "format": "int64"
                },
                "venta": {
                    "type": "integer",
                    "format": "int64"
                }
            }
        },
...
    }
}